refactor(frontend): move components from quiz to shared ui
